Changes to tests
