+++ libraries/botbuilder-dialog-prompts/test/package.json
@@ -9,6 +9,7 @@
   "author": "Michael Szul <michael@szul.us>",
   "license": "MIT",
   "dependencies": {
+    "@botbuildercommunity/dialog-prompts": "^0.1.0",
     "botbuilder": "^4.0.6",
     "botbuilder-dialogs": "^4.0.6",
     "restify": "^7.2.1"
